Move all the ContactUs framework logic outside the demo app
