+++ WordPress/ContactUs/Sources/WebView.swift
@@ -2,40 +2,10 @@ import SwiftUI
 import UIKit
 import WebKit
 
-public struct WebView: View {
+// This implementation is based upon
+// https://gist.github.com/joshbetz/2ff5922203240d4685d5bdb5ada79105
 
-    private let url: URL
-    @State private var alertPresented = false
-
-    public init(url: URL) {
-        self.url = url
-    }
-
-    public var body: some View {
-        VStack {
-            _WebView(url: url)
-            VStack(spacing: 4) {
-                Text("Can't find what you're looking for?").italic()
-                Button {
-                    self.alertPresented.toggle()
-                } label: {
-                    Text("Contact Support")
-                }
-                .alert(isPresented: $alertPresented) {
-                    Alert(
-                        title: Text("TODO"),
-                        message: Text("This should load the Zendesk flow"),
-                        dismissButton: .default(Text("Dismiss"))
-                    )
-                }
-            }
-        }
-    }
-}
-
-// Credits https://gist.github.com/joshbetz/2ff5922203240d4685d5bdb5ada79105
-
-struct _WebView: UIViewRepresentable {
+struct WebView: UIViewRepresentable {
 
     private let request: URLRequest
